refactor: React query custom hook으로 분리
